Share manifest.json and move manifest creation onto Project.
