+++ core/story-api-migration/src/main/java/org/lgna/project/io/ProjectIo.java
@@ -54,6 +54,8 @@ import java.io.OutputStream;
 
 public interface ProjectIo {
 
+  String MANIFEST_ENTRY_NAME = "manifest.json";
+
   String METADATA_ENTRY_NAME = "metadata.json";
   String VERSION_ENTRY_NAME = "version.txt";
